+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.1.0] - 2025-01-??
+## [0.1.1] - 2025-01-??
+### Added
+- Bus::get_eoi and Bus::clear_eoi functions
+- Removed BusCommand::trace_message
+
+### Changed
+
+### Fixed
+- Corrected bus state updates on Open - should be no state update
+
+## [0.1.0] - 2025-01-22
 ### Added
 - Initial release